+++ compiler-core/src/error.rs
@@ -2616,6 +2616,7 @@ a label or use a record constructor.",
                         location,
                         name,
                         hint,
+                        ..
                     } => {
                         let label_text = match hint {
                             UnknownTypeHint::AlternativeTypes(types) => did_you_mean(name, types),